Renamed lever_rss.py
I would like to create a small Flask app within this project. As a
result, I felt like the name didn't quite fit it's purpose.

Initially, I only saw this project as a script. However... my motives
have changed.

My current objective is to create a GitHub workflow that would run
the script once a day and have a Flask app with jobs.